Refactor to move threshold and expected detections into test config rather than guessed

+++ test/e2e/ids/xpack.go
@@ -311,7 +311,7 @@ type GetRecordsOptions struct {
 	End *time.Time
 	ExcludeInterim bool
 	Page *PageOptionsSpec
-	RecordScore float64
+	RecordScore int
 	Sort *string
 	Start *time.Time
 }
